Add MobileStepperExample component and fix nav item content
