+++ App.tsx
@@ -1,5 +1,5 @@
-import React, { useCallback, useState } from 'react';
-import { SafeAreaView, ScrollView, View, I18nManager, Text } from 'react-native';
+import React from 'react';
+import { SafeAreaView, ScrollView, View, I18nManager } from 'react-native';
 import { StackNavigationProp } from '@react-navigation/stack';
 import RNRestart from 'react-native-restart';
 import {
@@ -12,10 +12,8 @@ import {
     ListItemTag,
     ScoreCard,
     wrapIcon,
-    MobileStepper,
-    DotStepperVariant,
 } from '@pxblue/react-native-components';
-import { Card, Button, Divider } from 'react-native-paper';
+import { Card, Button } from 'react-native-paper';
 
 import MatIcon from 'react-native-vector-icons/MaterialIcons';
 import MaterialCommunityIcon from 'react-native-vector-icons/MaterialCommunityIcons';
@@ -26,6 +24,7 @@ import _Temp from '@pxblue/icons-svg/temp.svg';
 import _Humidity from '@pxblue/icons-svg/moisture.svg';
 import _Battery from '@pxblue/icons-svg/battery.svg';
 import { RootStackParamList } from './router';
+import { MobileStepperExample } from './components/mobile-stepper-example';
 
 const backgroundImage = require('./assets/images/farm.jpg');
 
@@ -68,50 +67,192 @@ type AppProps = {
     navigation: StackNavigationProp<RootStackParamList, 'App'>;
 };
 
-export const App: React.FC<AppProps> = ({ navigation }) => {
-    const totalSteps = 5;
-    const [currentStep, setCurrentStep] = useState(0);
-    const [mobileStepperVariant, setMobileStepperVariant] = useState<DotStepperVariant>('dots');
-
-    const updateStep = useCallback(
-        (delta: number): void => {
-            setCurrentStep(currentStep + delta);
-        },
-        [currentStep]
-    );
-
-    const cycleStepperVariant = (): void => {
-        switch (mobileStepperVariant) {
-            case 'dots':
-                setMobileStepperVariant('text');
-                break;
-            case 'text':
-                setMobileStepperVariant('progress');
-                break;
-            default:
-                setMobileStepperVariant('dots');
-                break;
-        }
-    };
-
-    return (
-        <View style={{ flex: 1, backgroundColor: PXBColors.gray[50] }}>
-            <Header
-                expandable={true}
-                title={'South Tin Mill'}
-                subtitle={'Gary Steel Works'}
-                info={'Online'}
-                navigation={{
-                    icon: MenuIcon,
+export const App: React.FC<AppProps> = ({ navigation }) => (
+    <View style={{ flex: 1, backgroundColor: PXBColors.gray[50] }}>
+        <Header
+            expandable={true}
+            title={'South Tin Mill'}
+            subtitle={'Gary Steel Works'}
+            info={'Online'}
+            navigation={{
+                icon: MenuIcon,
+                onPress: (): void => {
+                    navigation.openDrawer();
+                },
+            }}
+            actionItems={[
+                {
+                    icon: SwapIcon,
                     onPress: (): void => {
-                        navigation.openDrawer();
+                        toggleRTL();
                     },
+                },
+                {
+                    icon: MoreIcon,
+                    onPress: (): void => {
+                        /* do nothing */
+                    },
+                },
+            ]}
+            backgroundImage={backgroundImage}
+            searchableConfig={{ placeholder: 'Search', autoFocus: true }}
+        />
+        <ScrollView>
+            <Card style={{ padding: 0, margin: PADDING, marginBottom: 0 }}>
+                <EmptyState
+                    title={'Nothing Found'}
+                    description={'Not a single thing'}
+                    IconClass={ChartLineVariant}
+                    actions={
+                        <Button mode={'outlined'} labelStyle={{ color: PXBColors.blue[500] }}>
+                            Add a Device
+                        </Button>
+                    }
+                />
+            </Card>
+            <Card style={{ padding: 0, margin: PADDING, marginBottom: PADDING }}>
+                <HeroBanner>
+                    <Hero
+                        label={'Healthy'}
+                        value={96}
+                        units={'/100'}
+                        IconClass={A}
+                        fontSize={20}
+                        iconColor={PXBColors.green[500]}
+                    />
+                    <Hero label={'Battery'} value={'Full'} IconClass={Battery} iconColor={PXBColors.blue[500]} />
+                    <Hero label={'Estimated'} IconClass={Clock} iconColor={PXBColors.gray[500]}>
+                        <ChannelValue fontSize={20} value={1} units={'h'} />
+                        <ChannelValue fontSize={20} value={37} units={'m'} />
+                    </Hero>
+                    <Hero label={'Loaded'} IconClass={Pie} iconColor={PXBColors.blue[500]}>
+                        <ChannelValue fontSize={20} value={65} units={'%'} IconClass={ChartLineVariant} />
+                    </Hero>
+                    <Hero label={'Not Shown'} value={'5th Item'} IconClass={Battery} iconColor={PXBColors.blue[500]} />
+                </HeroBanner>
+                <InfoListItem
+                    divider={'full'}
+                    IconClass={Sunny}
+                    title={'Temperature'}
+                    rightComponent={<ChannelValue value={68} units={'°F'} />}
+                />
+                <InfoListItem
+                    divider={'full'}
+                    IconClass={Sunny}
+                    title={'Temperature'}
+                    rightComponent={<ChannelValue value={68} units={'°F'} />}
+                />
+                <InfoListItem
+                    divider={'full'}
+                    IconClass={Sunny}
+                    title={'Temperature'}
+                    rightComponent={
+                        <React.Fragment>
+                            <ChannelValue value={1} units={'h'} IconClass={Clock} />
+                            <ChannelValue value={24} units={'m'} />
+                        </React.Fragment>
+                    }
+                />
+            </Card>
+            <InfoListItem
+                title={'Emerson Field West'}
+                subtitle={['DG 100', 'EDR 5000', 'Online']}
+                statusColor={PXBColors.red[500]}
+                fontColor={PXBColors.red[500]}
+                hidePadding={true}
+                backgroundColor={'white'}
+                IconClass={Battery}
+                avatar
+                // dense
+                rightComponent={<MatIcon name={'mail'} size={24} color={PXBColors.black[500]} />}
+                divider={'partial'}
+                onPress={(): void => {
+                    /* do nothing */
                 }}
+            />
+            <InfoListItem
+                title={'South Hills Farm'}
+                subtitle={'Offline'}
+                backgroundColor={'white'}
+                statusColor={PXBColors.orange[500]}
+                divider={'full'}
+                IconClass={Pie}
+                iconAlign={'center'}
+                hidePadding={false}
+                fontColor={PXBColors.red[500]}
+                rightComponent={<ChannelValue value={15} units={'A'} />}
+                onPress={(): void => {
+                    /* do nothing */
+                }}
+            />
+            <InfoListItem
+                title={'Cherry East'}
+                backgroundColor={'white'}
+                subtitle={['DG 100', 'EDR 5000', 'Online']}
+                subtitleSeparator={'/'}
+                hidePadding
+                rightComponent={<ListItemTag label={'ACTIVE'} />}
+                onPress={(): void => {
+                    /* do nothing */
+                }}
+            />
+            <ScoreCard
+                style={{ margin: PADDING }}
+                headerTitle={'Portland Datacenter Long Name'}
+                headerSubtitle={'6 UPS Devices'}
+                headerInfo={'Attention Required'}
+                headerBackgroundImage={backgroundImage}
                 actionItems={[
                     {
-                        icon: SwapIcon,
+                        icon: MoreIcon,
                         onPress: (): void => {
-                            toggleRTL();
+                            /* do nothing */
+                        },
+                    },
+                ]}
+                badgeOffset={-55}
+                badge={
+                    <HeroBanner style={{ flex: 0, minWidth: 80, justifyContent: 'flex-end' }}>
+                        <Hero
+                            label={'Score'}
+                            iconSize={48}
+                            iconColor={PXBColors.green[500]}
+                            value={98}
+                            units={'/100'}
+                            IconClass={A}
+                        />
+                    </HeroBanner>
+                }
+                actionRow={
+                    <InfoListItem
+                        dense
+                        chevron
+                        hidePadding
+                        title={'View Location'}
+                        onPress={(): void => {
+                            /* do nothing */
+                        }}
+                    />
+                }
+            >
+                <View style={{ justifyContent: 'center', marginLeft: -16 }}>
+                    <InfoListItem dense IconClass={Notifications} iconColor={PXBColors.red[500]} title={'1 Alarm'} />
+                    <InfoListItem dense IconClass={Info} iconColor={PXBColors.blue[500]} title={'1 Event'} />
+                    <InfoListItem dense IconClass={Clouds} title={'Online'} />
+                </View>
+            </ScoreCard>
+            <ScoreCard
+                style={{ margin: PADDING, marginTop: 0 }}
+                headerColor={PXBColors.red[500]}
+                headerTitle={'Substation 3'}
+                headerSubtitle={'High Humidity Alarm'}
+                headerInfo={'5 Devices'}
+                headerBackgroundImage={backgroundImage}
+                actionItems={[
+                    {
+                        icon: MailIcon,
+                        onPress: (): void => {
+                            /* do nothing */
                         },
                     },
                     {
@@ -121,267 +262,47 @@ export const App: React.FC<AppProps> = ({ navigation }) => {
                         },
                     },
                 ]}
-                backgroundImage={backgroundImage}
-                searchableConfig={{ placeholder: 'Search', autoFocus: true }}
-            />
-            <ScrollView>
-                <Card style={{ padding: 0, margin: PADDING, marginBottom: 0 }}>
-                    <EmptyState
-                        title={'Nothing Found'}
-                        description={'Not a single thing'}
-                        IconClass={ChartLineVariant}
-                        actions={
-                            <Button mode={'outlined'} labelStyle={{ color: PXBColors.blue[500] }}>
-                                Add a Device
-                            </Button>
-                        }
-                    />
-                </Card>
-                <Card style={{ padding: 0, margin: PADDING, marginBottom: PADDING }}>
-                    <HeroBanner>
+                badge={
+                    <HeroBanner style={{ flex: 0, minWidth: 180, justifyContent: 'flex-end' }}>
                         <Hero
-                            label={'Healthy'}
-                            value={96}
-                            units={'/100'}
-                            IconClass={A}
-                            fontSize={20}
-                            iconColor={PXBColors.green[500]}
+                            label={'Temperature'}
+                            iconColor={PXBColors.black[500]}
+                            iconSize={70}
+                            value={69}
+                            units={'°F'}
+                            IconClass={Temp}
                         />
-                        <Hero label={'Battery'} value={'Full'} IconClass={Battery} iconColor={PXBColors.blue[500]} />
-                        <Hero label={'Estimated'} IconClass={Clock} iconColor={PXBColors.gray[500]}>
-                            <ChannelValue fontSize={20} value={1} units={'h'} />
-                            <ChannelValue fontSize={20} value={37} units={'m'} />
-                        </Hero>
-                        <Hero label={'Loaded'} IconClass={Pie} iconColor={PXBColors.blue[500]}>
-                            <ChannelValue fontSize={20} value={65} units={'%'} IconClass={ChartLineVariant} />
-                        </Hero>
                         <Hero
-                            label={'Not Shown'}
-                            value={'5th Item'}
-                            IconClass={Battery}
-                            iconColor={PXBColors.blue[500]}
+                            label={'Humidity'}
+                            iconSize={70}
+                            iconColor={PXBColors.blue[200]}
+                            value={78}
+                            units={'%'}
+                            IconClass={Humidity}
                         />
                     </HeroBanner>
+                }
+                actionRow={
                     <InfoListItem
-                        divider={'full'}
-                        IconClass={Sunny}
-                        title={'Temperature'}
-                        rightComponent={<ChannelValue value={68} units={'°F'} />}
+                        dense
+                        chevron
+                        hidePadding
+                        title={'View Location'}
+                        onPress={(): void => {
+                            /* do nothing */
+                        }}
                     />
-                    <InfoListItem
-                        divider={'full'}
-                        IconClass={Sunny}
-                        title={'Temperature'}
-                        rightComponent={<ChannelValue value={68} units={'°F'} />}
-                    />
-                    <InfoListItem
-                        divider={'full'}
-                        IconClass={Sunny}
-                        title={'Temperature'}
-                        rightComponent={
-                            <React.Fragment>
-                                <ChannelValue value={1} units={'h'} IconClass={Clock} />
-                                <ChannelValue value={24} units={'m'} />
-                            </React.Fragment>
-                        }
-                    />
-                </Card>
-                <InfoListItem
-                    title={'Emerson Field West'}
-                    subtitle={['DG 100', 'EDR 5000', 'Online']}
-                    statusColor={PXBColors.red[500]}
-                    fontColor={PXBColors.red[500]}
-                    hidePadding={true}
-                    backgroundColor={'white'}
-                    IconClass={Battery}
-                    avatar
-                    // dense
-                    rightComponent={<MatIcon name={'mail'} size={24} color={PXBColors.black[500]} />}
-                    divider={'partial'}
-                    onPress={(): void => {
-                        /* do nothing */
-                    }}
-                />
-                <InfoListItem
-                    title={'South Hills Farm'}
-                    subtitle={'Offline'}
-                    backgroundColor={'white'}
-                    statusColor={PXBColors.orange[500]}
-                    divider={'full'}
-                    IconClass={Pie}
-                    iconAlign={'center'}
-                    hidePadding={false}
-                    fontColor={PXBColors.red[500]}
-                    rightComponent={<ChannelValue value={15} units={'A'} />}
-                    onPress={(): void => {
-                        /* do nothing */
-                    }}
-                />
-                <InfoListItem
-                    title={'Cherry East'}
-                    backgroundColor={'white'}
-                    subtitle={['DG 100', 'EDR 5000', 'Online']}
-                    subtitleSeparator={'/'}
-                    hidePadding
-                    rightComponent={<ListItemTag label={'ACTIVE'} />}
-                    onPress={(): void => {
-                        /* do nothing */
-                    }}
-                />
-                <ScoreCard
-                    style={{ margin: PADDING }}
-                    headerTitle={'Portland Datacenter Long Name'}
-                    headerSubtitle={'6 UPS Devices'}
-                    headerInfo={'Attention Required'}
-                    headerBackgroundImage={backgroundImage}
-                    actionItems={[
-                        {
-                            icon: MoreIcon,
-                            onPress: (): void => {
-                                /* do nothing */
-                            },
-                        },
-                    ]}
-                    badgeOffset={-55}
-                    badge={
-                        <HeroBanner style={{ flex: 0, minWidth: 80, justifyContent: 'flex-end' }}>
-                            <Hero
-                                label={'Score'}
-                                iconSize={48}
-                                iconColor={PXBColors.green[500]}
-                                value={98}
-                                units={'/100'}
-                                IconClass={A}
-                            />
-                        </HeroBanner>
-                    }
-                    actionRow={
-                        <InfoListItem
-                            dense
-                            chevron
-                            hidePadding
-                            title={'View Location'}
-                            onPress={(): void => {
-                                /* do nothing */
-                            }}
-                        />
-                    }
-                >
-                    <View style={{ justifyContent: 'center', marginLeft: -16 }}>
-                        <InfoListItem
-                            dense
-                            IconClass={Notifications}
-                            iconColor={PXBColors.red[500]}
-                            title={'1 Alarm'}
-                        />
-                        <InfoListItem dense IconClass={Info} iconColor={PXBColors.blue[500]} title={'1 Event'} />
-                        <InfoListItem dense IconClass={Clouds} title={'Online'} />
-                    </View>
-                </ScoreCard>
-                <ScoreCard
-                    style={{ margin: PADDING, marginTop: 0 }}
-                    headerColor={PXBColors.red[500]}
-                    headerTitle={'Substation 3'}
-                    headerSubtitle={'High Humidity Alarm'}
-                    headerInfo={'5 Devices'}
-                    headerBackgroundImage={backgroundImage}
-                    actionItems={[
-                        {
-                            icon: MailIcon,
-                            onPress: (): void => {
-                                /* do nothing */
-                            },
-                        },
-                        {
-                            icon: MoreIcon,
-                            onPress: (): void => {
-                                /* do nothing */
-                            },
-                        },
-                    ]}
-                    badge={
-                        <HeroBanner style={{ flex: 0, minWidth: 180, justifyContent: 'flex-end' }}>
-                            <Hero
-                                label={'Temperature'}
-                                iconColor={PXBColors.black[500]}
-                                iconSize={70}
-                                value={69}
-                                units={'°F'}
-                                IconClass={Temp}
-                            />
-                            <Hero
-                                label={'Humidity'}
-                                iconSize={70}
-                                iconColor={PXBColors.blue[200]}
-                                value={78}
-                                units={'%'}
-                                IconClass={Humidity}
-                            />
-                        </HeroBanner>
-                    }
-                    actionRow={
-                        <InfoListItem
-                            dense
-                            chevron
-                            hidePadding
-                            title={'View Location'}
-                            onPress={(): void => {
-                                /* do nothing */
-                            }}
-                        />
-                    }
-                >
-                    <View style={{ justifyContent: 'center', marginLeft: -16 }}>
-                        <InfoListItem
-                            dense
-                            IconClass={Notifications}
-                            iconColor={PXBColors.red[500]}
-                            title={'1 Alarm'}
-                        />
-                        <InfoListItem dense IconClass={Info} iconColor={PXBColors.blue[500]} title={'1 Event'} />
-                        <InfoListItem dense IconClass={Clouds} title={'Online'} />
-                    </View>
-                </ScoreCard>
+                }
+            >
+                <View style={{ justifyContent: 'center', marginLeft: -16 }}>
+                    <InfoListItem dense IconClass={Notifications} iconColor={PXBColors.red[500]} title={'1 Alarm'} />
+                    <InfoListItem dense IconClass={Info} iconColor={PXBColors.blue[500]} title={'1 Event'} />
+                    <InfoListItem dense IconClass={Clouds} title={'Online'} />
+                </View>
+            </ScoreCard>
 
-                <Card style={{ padding: 0, margin: PADDING, marginBottom: PADDING }}>
-                    <Button mode="contained" onPress={(): void => cycleStepperVariant()}>
-                        Cycle Variant
-                    </Button>
-                    <Divider />
-                    <Text style={{ margin: 32, textAlign: 'center' }}>Step #{currentStep + 1}</Text>
-                    <Divider />
-                    <MobileStepper
-                        activeStep={currentStep}
-                        steps={totalSteps}
-                        leftButton={
-                            <View style={{ flex: 1 }}>
-                                <Button
-                                    style={{ width: 100, alignSelf: 'flex-start' }}
-                                    disabled={currentStep === 0}
-                                    onPress={(): void => updateStep(-1)}
-                                    mode="outlined"
-                                >
-                                    Back
-                                </Button>
-                            </View>
-                        }
-                        rightButton={
-                            <View style={{ flex: 1 }}>
-                                <Button
-                                    style={{ width: 100, alignSelf: 'flex-end' }}
-                                    disabled={currentStep === totalSteps - 1}
-                                    onPress={(): void => updateStep(1)}
-                                    mode="contained"
-                                >
-                                    Next
-                                </Button>
-                            </View>
-                        }
-                        variant={mobileStepperVariant}
-                    ></MobileStepper>
-                </Card>
-                <SafeAreaView />
-            </ScrollView>
-        </View>
-    );
-};
+            <MobileStepperExample />
+            <SafeAreaView />
+        </ScrollView>
+    </View>
+);
